setup node model
